Rename cop files to reflect renaming Primitives
