+++ lib/rubocop/cop/truffleruby/replace_with_primitive_class.rb
