+++ lib/rubocop/cop/truffleruby/replace_with_primitive_is_a.rb
