+++ spec/rubocop/cop/truffleruby/replace_with_primitive_class_spec.rb
