+++ spec/rubocop/cop/truffleruby/replace_with_primitive_equal_spec.rb
